Make clippy happy
