+++ src/cmd/mod.rs
@@ -6,4 +6,4 @@ pub use unit::Command as UnitCommand;
 
 pub(crate) type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
 
-pub(crate) const DEFAULT_URL: &'static str = "http://127.0.0.1:21812";
+pub(crate) const DEFAULT_URL: &str = "http://127.0.0.1:21812";
